refactor(imu): address CodeRabbit round 8 findings
- imu_task.h: add #include <stdint.h> for self-contained types; add
  @invariant to imu_task_stack_t documenting 2048-byte minimum
- telemetry.proto: reorder TelemetryData fields numerically (baro=12
  placed between fault_flags=11 and timestamp_us=14)
- rx_bmp280_regs.h: add @pre tags to bmp280_reg_t (POR sequence and
  I2C bus preconditions)
- rx_bno055_regs.h: add _Static_assert validating k_bno055_ms_per_tick==10
  matches TX_TIMER_TICKS_PER_SECOND=100 at compile time
- rx_bmp280.c: split bmp280_output_range_t into bmp280_temp_range_t
  (int32_t) and bmp280_press_range_t (uint32_t) for correct signedness
- test_rx_bmp280.c: fix dig_P1 doc value (36823->65410=0xFF82); add
  @warning about test-order dependency in main(); add _Static_assert and
  TEST_ASSERT postcondition checks to internal_load_* helpers
- telemetry_task.c: add second @pre and expand @details for
  internal_populate_imu_telemetry and internal_populate_baro_telemetry
- test_rx_bno055.c: add ordering comment before RUN_TEST calls; add
  @pre/@post tags to main() docblock
- rx_bno055.c: remove redundant (uint8_t) cast in internal_assemble_int16_le

+++ star-rx72n-firmware/libs/rx_bno055/src/rx_bno055.c
@@ -395,7 +395,7 @@ static inline int16_t internal_assemble_int16_le(uint8_t low, uint8_t high)
 {
   _Static_assert(k_bno055_shift_msb == 8U, "MSB shift must be 8 for little-endian assembly");
   _Static_assert(sizeof(uint16_t) * CHAR_BIT == 16U, "uint16_t must be 16 bits for assembly to be well-defined");
-  return (int16_t)((uint16_t)low | ((uint16_t)high << (uint8_t)k_bno055_shift_msb));
+  return (int16_t)((uint16_t)low | ((uint16_t)high << k_bno055_shift_msb));
 }
 
 /* =============================================================================
